mls_code_generator/src/requirements.txt
flask
flask_cors
waitress
PyYaml